+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 * Filter application details on data usage
 * Notify if application needs restart, thanks @[jpeg729](https://github.com/jpeg729)
 * Chain new uncaught exception handlers
+* Send usage data on service destroy
 * Updated Polish translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)